tests/wpt/tests/html/canvas/element/layers/2d.layer.layer-rendering-state-reset-in-layer.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.layer.layer-rendering-state-reset-in-layer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.layer-rendering-state-reset-in-layer</h1>
<p class="desc">Tests that layers ignore the global context filter.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Tests that layers ignore the global context filter.");
_addTest(function(canvas, ctx) {

  ctx.globalAlpha = 0.5;
  ctx.globalCompositeOperation = 'xor';
  ctx.shadowColor = '#0000ff';
  ctx.shadowOffsetX = 10;
  ctx.shadowOffsetY = 20;
  ctx.shadowBlur = 30;

  _assertSame(ctx.globalAlpha, 0.5, "ctx.globalAlpha", "0.5");
  _assertSame(ctx.globalCompositeOperation, 'xor', "ctx.globalCompositeOperation", "'xor'");
  _assertSame(ctx.shadowColor, '#0000ff', "ctx.shadowColor", "'#0000ff'");
  _assertSame(ctx.shadowOffsetX, 10, "ctx.shadowOffsetX", "10");
  _assertSame(ctx.shadowOffsetY, 20, "ctx.shadowOffsetY", "20");
  _assertSame(ctx.shadowBlur, 30, "ctx.shadowBlur", "30");

  ctx.beginLayer();

  _assertSame(ctx.globalAlpha, 1.0, "ctx.globalAlpha", "1.0");
  _assertSame(ctx.globalCompositeOperation, 'source-over', "ctx.globalCompositeOperation", "'source-over'");
  _assertSame(ctx.shadowColor, 'rgba(0, 0, 0, 0)', "ctx.shadowColor", "'rgba(0, 0, 0, 0)'");
  _assertSame(ctx.shadowOffsetX, 0, "ctx.shadowOffsetX", "0");
  _assertSame(ctx.shadowOffsetY, 0, "ctx.shadowOffsetY", "0");
  _assertSame(ctx.shadowBlur, 0, "ctx.shadowBlur", "0");

  ctx.endLayer();

  _assertSame(ctx.globalAlpha, 0.5, "ctx.globalAlpha", "0.5");
  _assertSame(ctx.globalCompositeOperation, 'xor', "ctx.globalCompositeOperation", "'xor'");
  _assertSame(ctx.shadowColor, '#0000ff', "ctx.shadowColor", "'#0000ff'");
  _assertSame(ctx.shadowOffsetX, 10, "ctx.shadowOffsetX", "10");
  _assertSame(ctx.shadowOffsetY, 20, "ctx.shadowOffsetY", "20");
  _assertSame(ctx.shadowBlur, 30, "ctx.shadowBlur", "30");

});
</script>
